# Conflicts:
#	README.md
#	src/main/webapp/WEB-INF/templates/member/login.jsp
#	src/main/webapp/static/css/login.css

+++ src/main/java/org/choongang/admin/sevices/config/MemberDeleteService.java
@@ -1,20 +1,30 @@
 package org.choongang.admin.sevices.config;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import org.choongang.board.exceptions.BoardNotFoundException;
 import org.choongang.global.config.annotations.Service;
-import org.choongang.member.entities.Member;
+import org.choongang.global.config.containers.BeanContainer;
+import org.choongang.global.exceptions.AlertException;
 import org.choongang.member.mappers.MemberMapper;
 
 @Service
 @RequiredArgsConstructor
 public class MemberDeleteService {
+
     private final MemberMapper mapper;
-    private final AllMemberConfigInfoService infoService;
 
-    public void delete(String email) {
-        Member member = infoService.get(email).orElseThrow(BoardNotFoundException::new);
-        mapper.delete(email);
+    public void process() {
+        HttpServletRequest request = BeanContainer.getInstance().getBean(HttpServletRequest.class);
 
+        String[] emails = request.getParameterValues("email");
+
+        if (emails == null || emails.length == 0) {
+            throw new AlertException("삭제할 회원을 선택하세요", HttpServletResponse.SC_BAD_REQUEST);
+        }
+
+        for (String email : emails) {
+            mapper.delete(email);
+        }
     }
 }
